+++ geo-types/CHANGES.md
@@ -1,9 +1,16 @@
 # Changes
 
-## Unrealeased
+## Unreleased
+
+* POSSIBLY BREAKING: Minimum supported version of Rust (MSRV) is now 1.70
+  * <https://github.com/georust/geo/pull/1134>
+
+## 0.7.12
 
 * Add `Polygon::try_exterior_mut` and `Polygon::try_interiors_mut`.
   <https://github.com/georust/geo/pull/1071>
+* Add `wkt!` macro to define geometries at compile time.
+  <https://github.com/georust/geo/pull/1063>
 
 ## 0.7.11
 * Bump rstar dependency